requirements.txt
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
certifi==2023.5.7
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646038465422/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cryptography @ file:///croot/cryptography_1677533068310/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
filelock @ file:///croot/filelock_1672387128942/work
fonttools==4.25.0
fsspec @ file:///croot/fsspec_1682525977151/work
huggingface-hub @ file:///croot/huggingface_hub_1684443994767/work
idna @ file:///croot/idna_1666125576474/work
importlib-metadata @ file:///croot/importlib-metadata_1678997070253/work
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1676919000169/work
joblib @ file:///croot/joblib_1685113087166/work
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
matplotlib @ file:///croot/matplotlib-suite_1679593461707/work
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
munkres==1.1.4
numexpr @ file:///croot/numexpr_1683221822650/work
numpy @ file:///work/mkl/numpy_and_numpy_base_1682953417311/work
opencv-python==4.7.0.72
packaging @ file:///croot/packaging_1678965309396/work
pandas==1.5.3
Pillow==9.4.0
ply==3.11
protobuf==3.20.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PyQt5-sip==12.11.0
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
PyYAML @ file:///croot/pyyaml_1670514731622/work
regex @ file:///tmp/abs_41f5bce5-0a2e-45aa-b231-1fd2fbd57753gfpe6sjm/croots/recipe/regex_1658257178822/work
requests @ file:///croot/requests_1682607517574/work
sacremoses @ file:///tmp/build/80754af9/sacremoses_1633107328213/work
scipy==1.10.1
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
tokenizers @ file:///opt/conda/conda-bld/tokenizers_1651822590771/work
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1604308577558/work
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1648827257044/work
tqdm @ file:///croot/tqdm_1679561862951/work
transformers==4.28.1
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
urllib3 @ file:///croot/urllib3_1686163155763/work
zipp @ file:///croot/zipp_1672387121353/work
